update readme and method names
